web_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.5.0
slowapi==0.1.8
redis==5.0.1
aiohttp>=3.9.0
gunicorn==21.2.0